renamed useFetchData.js -> useData.js, added usePostData() to useData.js for posting data to backend. added unimplmented handleSubmit() on TagInfo submission.
